Merge to Master

+++ frontend/src/components/Project.js
@@ -1,9 +1,11 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import Navbar from '../components/Navbar';
 import { Link } from 'react-router-dom';
 import { useForm } from 'react-hook-form';
 import { useAuth } from '../context/auth-context';
 import StarRating from '../components/StarRating';
+import StarRatingComment from '../components/StarRatingComment';
+import Swal from 'sweetalert2';
 import {
   addCommentProject,
   deleteCommentProject,
@@ -12,11 +14,23 @@ import {
   deleteProject
 } from '../http/projectService';
 
-export function Project({ project, comments, projectId, onUpdateProject, dispatch, assesment, assesmentAvg }) {
+export function Project({ project, comments, projectId, onUpdateProject, dispatch, assesment, assesmentC }) {
   const { handleSubmit, register, errors, formState } = useForm({
     mode: 'onBlur'
   });
   const { role, setRole, setUser } = useAuth();
+
+  const [files, setFiles] = useState([]);
+  const [previews, setPreviews] = useState([]);
+
+  useEffect(() => {
+    async function getPreviews() {
+      const promises = files.map(getPreview);
+      setPreviews(await Promise.all(promises));
+    }
+
+    getPreviews();
+  }, [files]);
 
   function refreshPage() {
     window.location.reload(false);
@@ -24,17 +38,33 @@ export function Project({ project, comments, projectId, onUpdateProject, dispatc
 
   function renderButtons(comment, index) {
     const actual_user = role === null ? null : role.userId;
-    if (comment.user === actual_user) {
-      return (
-        <div className="text-xs mt-2">
-          <button
-            onClick={() => deleteCommentProject(comment.id).then(refreshPage)}
-            className="bg-red-500 hover:bg-red-700 text-white font-bold py-1 px-1 border border-red-700 rounded mx-2"
-          >
-            <img src="https://img.icons8.com/android/15/000000/delete.png" />
-          </button>
-        </div>
-      );
+    if (role !== null) {
+      if (comment.user === actual_user) {
+        return (
+          <div className="text-xs mt-2">
+            <button
+              onClick={() => deleteCommentProject(comment.id).then(refreshPage)}
+              className="bg-red-500 hover:bg-red-700 text-white font-bold py-1 px-1 border border-red-700 rounded mx-2"
+            >
+              <img src="https://img.icons8.com/android/15/000000/delete.png" />
+            </button>
+          </div>
+        );
+      } else {
+        if (role.role === 'E') {
+          let assesmentForStarRating = 0;
+          for (const element of assesmentC) {
+            if (element.id === comment.id) {
+              assesmentForStarRating = element.type;
+            }
+          }
+          return (
+            <div className="pr-2">
+              <StarRatingComment comment={comment.id} assesment={assesmentForStarRating}></StarRatingComment>
+            </div>
+          );
+        }
+      }
     }
   }
 
@@ -44,8 +74,16 @@ export function Project({ project, comments, projectId, onUpdateProject, dispatc
       .then(() => window.location.reload())
       .catch((error) => {
         if (error.response.status === 401) {
-          setRole(null);
-          setUser(null);
+          localStorage.removeItem('currentUser');
+          Swal.fire({
+            icon: 'error',
+            title: 'Oops...',
+            text: 'Tú token de sesión ha expirado'
+          }).then(() => {
+            window.location.href = '/';
+          });
+        } else if (error.response.status === 400) {
+          window.location.href = '/404';
         }
       });
   };
@@ -55,10 +93,10 @@ export function Project({ project, comments, projectId, onUpdateProject, dispatc
       if (role.role === 'M') {
         return (
           <div className="w-full p-2">
-            <h1 className="font-bold mt-2">Nuevo comentario</h1>
+            <h1 className="font-bold text-copy-primary mt-2">Nuevo comentario</h1>
             <div className="w-full ">
               <form
-                className="bg-white md:shadow-md md:rounded pt-6 pb-8 mb-4 px-4 border-gray-200 border-2"
+                className="bg-background-comment border border-background-comment md:shadow-md md:rounded pt-6 pb-8 mb-4 px-4 "
                 onSubmit={handleSubmit(handleSend)}
                 noValidate
               >
@@ -71,7 +109,7 @@ export function Project({ project, comments, projectId, onUpdateProject, dispatc
                         value: 200
                       }
                     })}
-                    className="resize-none shadow appearance-none border rounded w-full py-2 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+                    className="resize-none shadow appearance-none bg-background-secondary border border-background-comment rounded w-full py-2 text-copy-primary leading-tight focus:outline-none focus:shadow-outline"
                     id="text"
                     rows="6"
                     type="text"
@@ -82,8 +120,12 @@ export function Project({ project, comments, projectId, onUpdateProject, dispatc
                 </div>
                 <div className="flex flex-wrap px-2 py-4 justify-between w-full">
                   <div className="flex flex-wrap align-bottom">
-                    <img class="w-10 h-10 rounded-full mr-4 self-center" src={role.avatar_url} alt="Avatar" />
-                    <p className="text-black text-sm leading-none self-center">{role.email}</p>
+                    <img
+                      class="w-10 h-10 rounded-full mr-4 object-cover self-center"
+                      src={role.avatar_url}
+                      alt="Avatar"
+                    />
+                    <p className="text-sm text-copy-primary leading-none self-center">{role.email}</p>
                   </div>
                   <div className="text-xs self-end mt-2">
                     <input
@@ -138,6 +180,19 @@ export function Project({ project, comments, projectId, onUpdateProject, dispatc
 
   function onChangeHandler(e) {
     setState({ ...estado, [e.target.name]: e.target.files[0] });
+    setFiles([...Array.from(e.target.files), ...files]);
+
+    const handleUpload = () => {
+      if (!files) {
+        return;
+      }
+
+      const data = new FormData();
+
+      files.forEach((file) => {
+        data.append('files', file);
+      });
+    };
   }
 
   const onSubmit = (formData, e) => {
@@ -152,6 +207,20 @@ export function Project({ project, comments, projectId, onUpdateProject, dispatc
         .then(() => (window.location.href = '/project/' + projectId))
         .then(() => {
           dispatch({ type: 'UPDATE_PROJECT', edit: 0 });
+        })
+        .catch((error) => {
+          if (error.response.status === 401) {
+            localStorage.removeItem('currentUser');
+            Swal.fire({
+              icon: 'error',
+              title: 'Oops...',
+              text: 'Tú token de sesión ha expirado'
+            }).then(() => {
+              window.location.href = '/';
+            });
+          } else if (error.response.status === 400) {
+            window.location.href = '/404';
+          }
         });
     } else {
       let promise1 = updateProject(projectId, formData);
@@ -159,6 +228,20 @@ export function Project({ project, comments, projectId, onUpdateProject, dispatc
         .then(() => (window.location.href = '/project/' + projectId))
         .then(() => {
           dispatch({ type: 'UPDATE_PROJECT', edit: 0 });
+        })
+        .catch((error) => {
+          if (error.response.status === 401) {
+            localStorage.removeItem('currentUser');
+            Swal.fire({
+              icon: 'error',
+              title: 'Oops...',
+              text: 'Tú token de sesión ha expirado'
+            }).then(() => {
+              window.location.href = '/';
+            });
+          } else if (error.response.status === 400) {
+            window.location.href = '/404';
+          }
         });
     }
   };
@@ -166,20 +249,23 @@ export function Project({ project, comments, projectId, onUpdateProject, dispatc
   function showComments() {
     if (comments.length !== 0) {
       return (
-        <div className="w-full flex flex-wrap justify-center border-gray-200 bg-gray-100 border-2 px-2 bg-white rounded mb-4 mt-10">
-          <h1 className="font-bold p-2">Comentarios más recientes</h1>
+        <div className="w-full flex flex-wrap justify-center border-background-borderf bg-background-borderf border-2 px-2 rounded mb-4 mt-10">
+          <h1 className="font-bold text-copy-primary p-2">Comentarios más recientes</h1>
           {comments.map((comment, index) => (
             <React.Fragment key={comment.id}>
               <hr className="style1 mb-2" />
-              <div className="break-all w-full bg-white border-gray-200 border-2 mx-2" id={index}>
-                <div className="flex flex-wrap py-4 justify-between w-full">
+              <div
+                className="mb-2 break-words w-full bg-background-comment rounded-lg border-background-borderf border-2 mx-2"
+                id={index}
+              >
+                <div className="flex flex-wrap py-4 justify-between w-full pl-2">
                   <div className="flex flex-wrap align-bottom">
                     <Link to={'/user/' + comment.user}>
-                      <img className="w-10 h-10 rounded-full ml-2 mr-2" src={comment.avatar_url} alt="Avatar" />
+                      <img className="w-10 h-10 rounded-full mr-2 object-cover" src={comment.avatar_url} alt="Avatar" />
                     </Link>
                     <div className="text-xs lg:text-sm self-center">
-                      <p className="text-black leading-none w-full">{comment.name + ' ' + comment.first_name}</p>
-                      <p className="text-gray-500">
+                      <p className="text-copy-primary leading-none w-full">{comment.name + ' ' + comment.first_name}</p>
+                      <p className="text-copy-primary">
                         {comment.updated_at === null
                           ? comment.created_at.replace('T', ' ').substring(0, 16)
                           : comment.updated_at.replace('T', ' ').substring(0, 16)}
@@ -190,9 +276,15 @@ export function Project({ project, comments, projectId, onUpdateProject, dispatc
                 </div>
                 <p
                   dangerouslySetInnerHTML={{ __html: comment.text.replace(/<br\s*\\?>/g, '\r\n') }}
-                  className="mb-2 text-xs lg:text-sm px-2"
+                  className="mb-2 text-xs text-copy-primary lg:text-sm px-2"
                   id={index + 'p'}
                 ></p>
+                <hr />
+                <h1 className="mb-2 text-copy-primary italic text-xs lg:text-sm px-2 mt-2">
+                  {(comment.avg === null ? '' : 'Valorado ' + Math.round(comment.avg * 100) / 100) +
+                    (comment.avg === null ? 'Sin valoración todavía' : ' / ') +
+                    (comment.avg === null ? '' : comment.counter + ' votos')}
+                </h1>
               </div>
             </React.Fragment>
           ))}
@@ -201,7 +293,7 @@ export function Project({ project, comments, projectId, onUpdateProject, dispatc
       );
     } else {
       return (
-        <div className="w-full flex flex-wrap justify-center border-gray-200 bg-gray-100 border-2 px-2 bg-white rounded mb-4 mt-10">
+        <div className="w-full flex flex-wrap justify-center text-copy-primary border-background-borderf bg-background-borderf border-2 px-2 rounded mb-4 mt-10">
           <h1 className="font-bold p-2">No hay comentarios en este proyecto</h1>
           {renderNewComment()}
         </div>
@@ -210,12 +302,15 @@ export function Project({ project, comments, projectId, onUpdateProject, dispatc
   }
 
   function showStarIfLogged() {
+    const actual_user = role === null ? null : role.userId;
     if (role) {
-      return (
-        <div className="mt-2">
-          <StarRating assesment={assesment} project={projectId}></StarRating>
-        </div>
-      );
+      if (actual_user !== project[0].user) {
+        return (
+          <div className="mt-2">
+            <StarRating assesment={assesment} project={projectId}></StarRating>
+          </div>
+        );
+      }
     }
   }
   function something(onUpdateProject, projectC) {
@@ -230,10 +325,12 @@ export function Project({ project, comments, projectId, onUpdateProject, dispatc
             <div className="w-full md:w-3/5">
               {project.map((project, index) => (
                 <div key={project.id} className="break-words rounded mx-4 md:mx-0">
-                  <div className="p-2 text-center font-serif font-bold text-xl lg:text-4xl tracking-wide ">
+                  <div className="p-2 text-center font-serif text-copy-primary font-bold text-xl lg:text-4xl tracking-wide ">
                     {project.title}
                   </div>
-                  <div className="text-center font-serif text:l lg:text-xl tracking-wide ">{project.subtitle}</div>
+                  <div className="text-center font-serif text:l text-copy-primary lg:text-xl tracking-wide ">
+                    {project.subtitle}
+                  </div>
                   <div className="flex flex-wrap mt-4 md:mt-10">
                     <div className="lg:w-1/2 -mx-4 md:mx-0">
                       <div class="bg-white sm:rounded-lg overflow-hidden">
@@ -242,28 +339,29 @@ export function Project({ project, comments, projectId, onUpdateProject, dispatc
                     </div>
                     <div className="order-2 lg:order-3">
                       {showStarIfLogged()}
-                      <div className="w-full">
-                        <div className="text-black font-semibold text-sm w-full">
+                      <div className="w-full mt-2">
+                        <div className="text-copy-primary font-semibold text-sm w-full">
                           Valoración media:{' '}
-                          {(assesmentAvg === null ? '' : assesmentAvg.avg.substring(0, 4)) +
-                            (assesmentAvg === null ? '0' : ' / ') +
-                            (assesmentAvg === null ? '' : assesmentAvg.counter) +
-                            ' votos'}
+                          {project.avg === null
+                            ? 'Sin valoración todavía'
+                            : Math.round(project.avg * 100) / 100 + ' / ' + project.counter + ' opiniones'}
                         </div>
-                        <div className="text-gray-600 font-semibold text-sm w-full">Ubicación: {project.ubication}</div>
-                        <div className="text-gray-600 font-semibold text-sm mb-2 w-full">
+                        <div className="text-copy-primary font-semibold text-sm w-full">
+                          Ubicación: {project.ubication}
+                        </div>
+                        <div className="text-copy-primary font-semibold text-sm mb-2 w-full">
                           Categoría: {project.category}
                         </div>
                       </div>
-                      <h1 className="font-bold w-full mt-6">Descripción del Proyecto </h1>
+                      <h1 className="font-bold text-copy-primary text-lg w-full mt-6">Descripción del Proyecto </h1>
                       <p
                         dangerouslySetInnerHTML={{ __html: project.text }}
-                        className="break-all text-gray-700 text-lg"
+                        className="break-words text-gray-600 text-lg"
                       ></p>
                     </div>
                     <div className="w-full lg:w-1/2 md:p-16 mt-10 lg:mt-0 order-3 lg:order-2">
                       <div className="border-grey-400">
-                        <h1 className="text-black font-semibold text-2xl text-center">Perfil del creador</h1>
+                        <h1 className="font-semibold text-copy-primary text-2xl text-center">Perfil del creador</h1>
                         <div className="flex flex-wrap justify-center mt-4">
                           <Link
                             to={'/user/' + project.user}
@@ -271,7 +369,7 @@ export function Project({ project, comments, projectId, onUpdateProject, dispatc
                           >
                             <img className="h-full w-full object-cover" src={project.avatar_url} alt="Your avatar" />
                           </Link>
-                          <div className="text-black font-semibold text-center break-all p-4 w-full">
+                          <div className="text-copy-primary font-semibold text-center break-all p-4 w-full">
                             {project.name + ' ' + project.first_name}
                           </div>
                           <div
@@ -279,7 +377,7 @@ export function Project({ project, comments, projectId, onUpdateProject, dispatc
                               __html:
                                 project.description === null ? '' : project.description.replace(/<br\s*\\?>/g, '\r\n')
                             }}
-                            className="h-6 text-black font-semibold text-center truncate text-sm w-full"
+                            className="h-6 text-copy-primary font-semibold text-center truncate text-sm w-full"
                           ></div>
                           <a
                             href={'/user/' + project.user}
@@ -441,6 +539,17 @@ export function Project({ project, comments, projectId, onUpdateProject, dispatc
                     className="shadow appearance-none border rounded py-2 px-1 mb-2 text-gray-700 w-full leading-tight focus:outline-none focus:shadow-outline"
                     onChange={onChangeHandler}
                   />
+                  <ul>
+                    {files.map((file, i) => (
+                      <li key={file.name}>
+                        <img src={previews[i]} alt={file.name} />
+                        <span>{file.name}</span>
+                      </li>
+                    ))}
+                  </ul>
+
+                  <label className="block text-gray-700 text-sm font-bold">Documentación del proyecto</label>
+
                   <div className="mt-2">
                     <button
                       className="bg-blue-500 text-white font-bold py-2 mr-2 px-2 rounded focus:outline-none focus:shadow-outline"
@@ -465,4 +574,19 @@ export function Project({ project, comments, projectId, onUpdateProject, dispatc
   }
 
   return <div>{something(onUpdateProject, project[0])}</div>;
+}
+function getPreview(file) {
+  return new Promise((resolve) => {
+    if (file && file.type.includes('image')) {
+      let reader = new FileReader();
+
+      reader.onloadend = function() {
+        resolve(reader.result);
+      };
+
+      reader.readAsDataURL(file);
+    } else {
+      resolve('http://via.placeholder.com/50x50');
+    }
+  });
 }
